refactor: moved datetime_utils
